+++ src/Core/Contracts/Events/EventInterface.php
@@ -11,14 +11,9 @@
 
 declare(strict_types=1);
 
-namespace Bitrix24\SDK\Application\Requests\Events;
+namespace Bitrix24\SDK\Core\Contracts\Events;
 
-
-use Bitrix24\SDK\Application\ApplicationStatus;
-use Bitrix24\SDK\Core\Credentials\Scope;
-use Bitrix24\SDK\Core\Exceptions\InvalidArgumentException;
-use Bitrix24\SDK\Core\Exceptions\UnknownScopeCodeException;
-use Bitrix24\SDK\Core\Result\AbstractItem;
+use Bitrix24\SDK\Application\Requests\Events\EventAuthItem;
 
 interface EventInterface
 {